requirements.txt
crewai==0.28.8
crewai_tools==0.1.6 
langchain_community  
langchain_openai 
qdrant-client 
pdfplumber 
langchainhub
python-dotenv